+++ apps/events-helsinki/src/domain/search/eventSearch/advancedSearch/__tests__/Search.test.tsx
